Added Default wallet clarification in Guide tab
The default wallet needs to be explicitly loaded if you're running the Electron wallet. (And similar for the game and vault wallets if you're running the QT wallet.) Additional information was added to the Guide tab for this, i.e. either load the wallet or delete the default wallet setting.