Added brute force password recovery
Added a very basic brute force password recovery. It requires a dictionary file, and isn't super fast, but it's a glimmer of hope for anyone that has lost their password, but can remember enough to create a dictionary of possible passwords.

The code is heavily commented for anyone that wants to improve on it.